fix(ui): model branch permission editability
