+++ apps/agor-ui/src/components/BranchModal/BranchModal.tsx
@@ -223,6 +223,8 @@ export const BranchModal: React.FC<BranchModalProps> = ({
                 canEdit={form.canEditPermissions}
                 allUsers={form.allUsers}
                 allGroups={form.allGroups}
+                groupGrantsStatus={form.groupGrantsStatus}
+                groupGrantsError={form.groupGrantsError}
                 currentUser={currentUser}
                 state={form.permissions}
                 setField={form.setPermissions}
