+++ apps/agor-ui/src/components/BranchModal/useBranchModalForm.test.tsx
@@ -19,143 +19,10 @@
  *   6. RBAC-disabled instances don't trip permissionsChanged.
  */
 
-import type { AgorClient, AssistantConfig, Branch, User } from '@agor-live/client';
 import { act, renderHook, waitFor } from '@testing-library/react';
-import { App as AntApp } from 'antd';
-import type { ReactNode } from 'react';
 import { describe, expect, it } from 'vitest';
+import { makeAssistantBranch, makeBranch, makeStubClient, makeUser, wrapper } from './testUtils';
 import { useBranchModalForm } from './useBranchModalForm';
-
-function wrapper({ children }: { children: ReactNode }) {
-  return <AntApp>{children}</AntApp>;
-}
-
-interface ServiceCall {
-  service: string;
-  method: 'find' | 'create' | 'patch' | 'remove' | 'findAll';
-  args: unknown[];
-}
-
-interface StubOptions {
-  owners?: User[];
-  users?: User[];
-  rbac404?: boolean;
-  groupGrants404?: boolean;
-  failBranchPatch?: boolean;
-  /** Throw a 500-style error on the initial owners.find load. */
-  failOwnersFind?: boolean;
-}
-
-function makeStubClient(opts: StubOptions = {}): { client: AgorClient; calls: ServiceCall[] } {
-  const owners = [...(opts.owners ?? [])];
-  const users = opts.users ?? [];
-  const calls: ServiceCall[] = [];
-
-  const client = {
-    service(path: string) {
-      return {
-        async find(args: unknown) {
-          calls.push({ service: path, method: 'find', args: [args] });
-          if (path === 'branches/:id/owners') {
-            if (opts.rbac404) {
-              const err = new Error('not found') as Error & { code?: number };
-              err.code = 404;
-              throw err;
-            }
-            if (opts.failOwnersFind) {
-              const err = new Error('database is down') as Error & { code?: number };
-              err.code = 500;
-              throw err;
-            }
-            return owners;
-          }
-          if (path === 'branches/:id/group-grants' && opts.groupGrants404) {
-            const err = new Error('not found') as Error & { code?: number };
-            err.code = 404;
-            throw err;
-          }
-          return [];
-        },
-        async findAll(args: unknown) {
-          calls.push({ service: path, method: 'findAll', args: [args] });
-          if (path === 'users') return users;
-          return [];
-        },
-        async create(body: unknown, params?: unknown) {
-          calls.push({ service: path, method: 'create', args: [body, params] });
-          if (path === 'branches/:id/owners') {
-            const userId = (body as { user_id: string }).user_id;
-            const newUser = users.find((u) => u.user_id === userId);
-            if (newUser) owners.push(newUser);
-            return newUser ?? { user_id: userId };
-          }
-          return body;
-        },
-        async patch(id: string, body: unknown, params?: unknown) {
-          calls.push({ service: path, method: 'patch', args: [id, body, params] });
-          if (path === 'branches' && opts.failBranchPatch) {
-            throw new Error('daemon exploded');
-          }
-          return { ...(body as object), branch_id: id };
-        },
-        async remove(id: string, params?: unknown) {
-          calls.push({ service: path, method: 'remove', args: [id, params] });
-          if (path === 'branches/:id/owners') {
-            const idx = owners.findIndex((o) => o.user_id === id);
-            if (idx >= 0) owners.splice(idx, 1);
-          }
-          return { user_id: id };
-        },
-      };
-    },
-  } as unknown as AgorClient;
-
-  return { client, calls };
-}
-
-function makeUser(overrides: Partial<User> = {}): User {
-  return {
-    user_id: 'user-1',
-    email: 'alice@example.com',
-    role: 'admin',
-    ...overrides,
-  } as unknown as User;
-}
-
-function makeBranch(overrides: Partial<Branch> = {}): Branch {
-  return {
-    branch_id: 'branch-1',
-    name: 'feature/foo',
-    repo_id: 'repo-1',
-    board_id: undefined,
-    issue_url: undefined,
-    pull_request_url: undefined,
-    notes: '',
-    mcp_server_ids: [],
-    others_can: 'session',
-    others_fs_access: 'read',
-    dangerously_allow_session_sharing: false,
-    ...overrides,
-  } as unknown as Branch;
-}
-
-function makeAssistantBranch(
-  overrides: Partial<Branch> = {},
-  configOverrides: Partial<AssistantConfig> = {}
-): Branch {
-  return makeBranch({
-    board_id: 'board-1',
-    custom_context: {
-      assistant: {
-        kind: 'assistant',
-        displayName: 'My Assistant',
-        emoji: '🤖',
-        ...configOverrides,
-      } as AssistantConfig,
-    },
-    ...overrides,
-  });
-}
 
 describe('useBranchModalForm — unified save', () => {
   it('sends ONE branches.patch combining general + permission fields, plus owners diffs', async () => {
@@ -507,9 +374,11 @@ describe('useBranchModalForm — unified save', () => {
     expect(result.current.rbacEnabled).toBe(true);
     expect(result.current.canViewPermissions).toBe(true);
     expect(result.current.canEditPermissions).toBe(true);
+    expect(result.current.groupGrantsStatus).toBe('unavailable');
+    expect(result.current.groupGrantsError?.message).toBe('not found');
   });
 
-  it('allows admins to view permissions even when owner rows are incomplete', async () => {
+  it('allows admins to view but not edit permissions when owner rows are incomplete', async () => {
     const alice = makeUser({ user_id: 'user-1', email: 'alice@example.com', role: 'admin' });
     const branch = makeBranch();
     const { client } = makeStubClient({ owners: [], users: [alice] });
@@ -522,7 +391,7 @@ describe('useBranchModalForm — unified save', () => {
     await waitFor(() => expect(result.current.loadingOwners).toBe(false));
 
     expect(result.current.canViewPermissions).toBe(true);
-    expect(result.current.canEditPermissions).toBe(true);
+    expect(result.current.canEditPermissions).toBe(false);
   });
 
   it('hides permissions from non-admin users who are not branch owners after owners load', async () => {
